rename Apologies module to trunk
IntelliJ IDEA wouldn't let me name the project "Apologies". It kept thinking the project was named "trunk" because that's the parent folder name. It kept trying to create a new module called trunk whenever I opened the project. So I'm giving up on having IDEA call it "Apologies" and letting it be called "trunk" to avoid annoying errors.