Add slides to board image
I also checked in the Photoshop file I used to generate it, so it’s easier to edit later.

The slide colors are the same as the safe zone colors, except for the start triangle and end circle, which are 10 L (Lightness in Photoshop’s color model) darker than the normal color.