Add the design document we submitted earlier
We submitted it on March 13, 2014 for “Homework 4: Design Document”.

This design document was more to get something to submit than to actually plan out a good design. The design specified here is not a great one, and future maintainers should probably use their own design skills to work on the code, and ignore this.